backend/requirements/prod.txt
asgiref==3.3.4
beautifulsoup4==4.9.3
boto3==1.17.78
botocore==1.20.78
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docopt==0.6.2
idna==2.10
jmespath==0.10.0
Js2Py==0.71
packaging==20.9
pipwin==0.5.1
PyAudio==0.2.11
pyjsparser==2.7.1
PyJWT==1.7.1
pyparsing==2.4.7
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
s3transfer==0.4.2
six==1.16.0
soupsieve==2.2.1
SpeechRecognition==3.8.1
sqlparse==0.4.1
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
